Fix overflow of map buttons on small devices
